[REFAC: #37] - Criando uma camada de serviço.
close #37
