+++ pyheat1d_web/core/views.py
@@ -2,7 +2,6 @@ import json
 from functools import partial
 from pathlib import Path
 
-from django.conf import settings
 from django.contrib import messages
 from django.http import HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, render, resolve_url
@@ -11,36 +10,7 @@ from pyheat1d.singleton import Singleton
 
 from .forms import EditSimulationForm, NewSimulationForm
 from .models import Simulation
-
-
-def _get_simulations_base_folder():
-    return Path(settings.MEDIA_ROOT)
-
-
-def _create_or_update_simulation_case(new_case, indent=2, update=False):
-    bcs = {
-        "lbc": {"type": 1, "params": {"value": new_case.pop("lbc_value")}},
-        "rbc": {"type": 1, "params": {"value": new_case.pop("rbc_value")}},
-    }
-
-    props = {"k": 1.0, "ro": 1.0, "cp": 1.0}
-
-    new_case.update(bcs)
-    new_case.update({"prop": props, "write_every_steps": 100})
-
-    tag = new_case.pop("tag")
-    base_folder = _get_simulations_base_folder()
-    if not update:
-        simulation_folder = base_folder / tag
-        if not simulation_folder.exists():
-            simulation_folder.mkdir(parents=True)
-
-    case_file = base_folder / f"{tag}/case.json"
-
-    # TODO: trata a exceção
-    json.dump(new_case, case_file.open(mode="w"), indent=indent)
-
-    return case_file
+from .services import create_or_update_simulation_case, delete_simulation_folder
 
 
 def create_simulation_form(request):
@@ -54,7 +24,7 @@ def create_simulation_form(request):
                 context={"form": form},
             )
 
-        form.instance.input_file = _create_or_update_simulation_case(form.cleaned_data.copy())
+        form.instance.input_file = create_or_update_simulation_case(form.cleaned_data.copy())
 
         form.save()
 
@@ -121,23 +91,6 @@ def detail_simulation(request, pk):
     return render(request, "core/detail_simulation.html", context=context)
 
 
-def _delete_simulation_folder(input_file):
-    if input_file.exists():
-        input_file.unlink()
-
-    base_dir = input_file.parent
-
-    mesh_file = base_dir / "mesh.json"
-    if mesh_file.exists():
-        mesh_file.unlink()
-
-    results_file = base_dir / "results.json"
-    if results_file.exists():
-        results_file.unlink()
-
-    base_dir.rmdir()
-
-
 def delete_simulation(request, pk):
     url_out = resolve_url("core:list_simulation")
 
@@ -148,7 +101,7 @@ def delete_simulation(request, pk):
         return HttpResponseRedirect(url_out)
 
     try:
-        _delete_simulation_folder(Path(sim.input_file))
+        delete_simulation_folder(Path(sim.input_file))
     except OSError:  # TODO: Criar um exeção personalizadas
         messages.error(request, f"Não foi possivel deletar o diretório da Simulação {sim.tag}.")
         return HttpResponseRedirect(url_out)
@@ -208,7 +161,7 @@ def edit_simulation_form(request, pk):
         form.save()
 
         case_data = {**form.cleaned_data.copy(), "tag": sim.tag}
-        _create_or_update_simulation_case(case_data, update=True)
+        create_or_update_simulation_case(case_data, update=True)
         messages.success(request, f"Dados da simulação atualizados {sim.tag}")
         return HttpResponseRedirect(resolve_url("core:list_simulation"))
     else:
